backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "ffmpeg-python>=0.2.0",
    "mlx-lm>=0.21.4",
    "openai-whisper>=20240930",
    "opencv-python>=4.11.0.86",
    "python-multipart>=0.0.20",
    "sentence-transformers>=3.4.1",
    "ultralytics>=8.3.77",
    "uvicorn>=0.34.0",
    "websockets>=15.0",
]
